Add sixel_decode_raw_rgba, fix decoding of high-color SIXEL's
Closes #15.
Closes #16.
Closes saitoha#149.
Closes saitoha#28.
Closes saitoha#61.

This PR fixes both problems at once. It does so by adding a new decode
function, and using it by default in `img2sixel`.

This function returns to the user RGBA data instead of palette'd data.
This sidesteps the many issues I found when considering raising the
maximum size of a palette to 255*255*255*255. For one thing, I already
knew I needed at some point for Kitty project something like this so I
wouldn't need to allocate my own, which is needless effort when the
decoder can do it.

This is ready for review. @dankamongmen may have time. I'll leave it for
a while as I go back to working on Kitty. If no one has time to review,
I'll merge it in ~4 days or so.