correctly use feature options
In every case, feature options are passed to the required: kwarg, which
means that:
- if enabled, failure to find the dependency results in immediate error
- if auto, try to find the dependency but disable it if not found
- if disabled, log a message that the dependency was skipped

It is not useful to then further check if the feature was enabled,
before adding it into the built software. The only difference this ever
makes is in the "auto" case, which then completely ignores that the user
specified "auto", discards any found dependencies, and acts as though
"disabled" was specified.

Allow the feature options to fulfill their specified role.