meson: add a declared dependency interface for the libsixel library
This allows third-party consumers of libsixel which also use meson, to
add a subproject wrap and link to libsixel_dep the same way they can
link to dependency('libsixel').

Although generally using pkg-config and system libraries is preferred
meson subproject wraps can be used in fallback mode, and
dependency('libsixel') will first check for a system version, and then
if it cannot find one, it will check if a subproject defines one
instead, but the subproject version needs to also expose the correct
include directory using the declared dependency interface.